01a/pyproject.toml
[project]
name = "01a"
version = "0.1.0"
description = "A project to parse data files"
authors = [
    {name = "HaDoCK", email = "jako7096@stud.kea.dk"}
]
requires-python = ">=3.12"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "pyyaml (>=6.0.2)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:app"

[[tool.poetry.packages]]
include = "src"